requirements.txt
flask
flask-bootstrap5
flask-wtf
wtforms
pandas
numpy
scikit-learn
tensorflow
joblib
gunicorn
matplotlib